# Swagger_Documentation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee Management API",
    "description": "APIs for managing employees",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/leaves/{id}/reject": {
      "put": {
        "tags": [
          "Leave"
        ],
        "summary": "Reject a leave request",
        "description": "Rejects the leave request for the employee identified by the provided leave request ID.",
        "operationId": "rejectLeave",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the leave request to reject",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leave request rejected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveDTO"
                }
              }
            }
          },
          "404": {
            "description": "Leave request not found for the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/leaves/{id}/approve": {
      "put": {
        "tags": [
          "Leave"
        ],
        "summary": "Approve a leave request",
        "description": "Approves the leave request for the employee identified by the provided leave request ID.",
        "operationId": "approveLeave",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the leave request to approve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leave request approved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveDTO"
                }
              }
            }
          },
          "404": {
            "description": "Leave request not found for the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/leaves": {
      "post": {
        "tags": [
          "Leave"
        ],
        "summary": "Submit a new leave request",
        "description": "Allows employees to submit a new leave request with the provided leave details such as start date, end date, and reason for leave.",
        "operationId": "applyForLeave",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeaveDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Leave request submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid leave data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found with the provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/leaves/{departmentId}": {
      "get": {
        "tags": [
          "Leave"
        ],
        "summary": "Get all leave requests by department",
        "description": "Retrieves all leave requests submitted by employees within a specific department, identified by the department ID.",
        "operationId": "getAllLeavesByDepartmentId",
        "parameters": [
          {
            "name": "departmentId",
            "in": "path",
            "description": "ID of the department to retrieve leave requests for",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of leave requests retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveDTO"
                }
              }
            }
          },
          "404": {
            "description": "Department not found for the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LeaveDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "employeeName": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          }
        }
      },
      "ErrorResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "CreateLeaveDTO": {
        "type": "object",
        "description": "Details of the leave request including dates and reason",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int64"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "employeeId",
          "endDate",
          "reason",
          "startDate"
        ]
      },
...
    }
  }
}